Include @quatrix's reconnect logic
This functionality was originally proposed and implemented by @quatrix
via https://github.com/joshgummersall/statsd-riemann-backend/pull/11.
Due to some unfortunate timing around a rewrite of the codebase in
ES2015, it was easier to include the functionality by simply rewriting
it rather than including the commits directly. For all intents and
purposes this commit should be attributed to @quatrix.

@quatrix: https://github.com/quatrix